extract BenchmarkInfo to a separate file
