+++ flutter/lib/benchmark/info.dart
@@ -1,0 +1,59 @@
+import 'package:flutter_svg/svg.dart';
+
+import 'package:mlperfbench/icons.dart';
+import 'package:mlperfbench/info.dart';
+import 'package:mlperfbench/protos/mlperf_task.pb.dart' as pb;
+
+enum BenchmarkTypeEnum {
+  unknown,
+  imageClassification,
+  objectDetection,
+  imageSegmentation,
+  languageUnderstanding,
+}
+
+class BenchmarkInfo {
+  final pb.ModelConfig modelConfig;
+
+  /// 'Object Detection', 'Image Classification (offline)', and so on.
+  final String taskName;
+
+  BenchmarkInfo(this.modelConfig, this.taskName);
+
+  String get name => modelConfig.name;
+
+  bool get isOffline => modelConfig.scenario == 'Offline';
+
+  double get maxScore => MAX_SCORE[modelConfig.id]!;
+
+  /// 'IC', 'OD', and so on.
+  String get code => modelConfig.id.split('_').first;
+
+  /// 'SingleStream' or 'Offline'.
+  String get scenario => modelConfig.scenario;
+
+  BenchmarkTypeEnum get type => _typeFromCode();
+
+  SvgPicture get icon => BENCHMARK_ICONS[scenario]?[code] ?? Icons.logo;
+
+  SvgPicture get iconWhite =>
+      BENCHMARK_ICONS_WHITE[scenario]?[code] ?? Icons.logo;
+
+  @override
+  String toString() => 'Benchmark:${modelConfig.id}';
+
+  BenchmarkTypeEnum _typeFromCode() {
+    switch (code) {
+      case 'IC':
+        return BenchmarkTypeEnum.imageClassification;
+      case 'OD':
+        return BenchmarkTypeEnum.objectDetection;
+      case 'IS':
+        return BenchmarkTypeEnum.imageSegmentation;
+      case 'LU':
+        return BenchmarkTypeEnum.languageUnderstanding;
+      default:
+        return BenchmarkTypeEnum.unknown;
+    }
+  }
+}
